pyproject.toml
[project]
name = "cron-cli-scheduler"
version = "0.1.0"
description = "Cross-platform cron-like scheduler with task management and MCP integration"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Cron Scheduler Team"}
]
keywords = ["cron", "scheduler", "task", "mcp", "automation", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
]

dependencies = [
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "croniter>=1.0.0",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0.0",
    "starlette>=0.27.0",
    "uvicorn>=0.20.0",
    "colorama>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
cron-cli = "scheduler.cli:main"
cron-scheduler-daemon = "scheduler.daemon:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scheduler"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D"]
ignore = ["D100", "D104"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
